requirements.txt
aiofiles>=23.1.0
bcrypt>=4.0.0
falcon>=3.1.0
gunicorn>=21.2.0
spectree>=1.2.0
pydantic>=2.1.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.19
alembic>=1.11.0
sqlalchemy-utils>=0.41.0
pytest>=7.4.0

# if errors, try --use-pep517 option
psycopg2>=2.9.5